refactor: make create user factory and create request refactor
