+++ src/http/controllers/user-controller.ts
@@ -1,8 +1,6 @@
 import { NextFunction, Request, Response } from "express";
-
 import z from "zod";
 import { AppError } from "../middlewares/error-handler";
-import { hash } from "bcryptjs";
 import { User } from "../../@types/user";
 import prisma from "../../lib/prisma";
 import { SuccessResponse } from "../../@types/response";
@@ -56,81 +54,6 @@ export const getUserById = async (
 
     res.status(200).json(response);
   } catch (error) {
-    if (error instanceof z.ZodError) {
-      const validationErrors = error.errors.map((err) => ({
-        field: err.path.join("."),
-        message: err.message,
-      }));
-
-      const validationError = new AppError(
-        "Invalid request data",
-        400,
-        validationErrors // Passando os detalhes como terceiro parâmetro
-      );
-
-      return next(validationError);
-    }
-
-    // Se for um erro customizado, passa direto
-    if (error instanceof AppError) {
-      return next(error);
-    }
-
-    // Para outros erros, cria um erro genérico
-    const genericError = new AppError(
-      error instanceof Error ? error.message : "Internal Server Error",
-      500
-    );
-    next(genericError);
-  }
-};
-
-// POST /api/users - Criar novo usuário
-export const createUser = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    const createUserBodySchema = z.object({
-      name: z.string(),
-      email: z.string().email(),
-      password: z.string().min(6),
-    });
-
-    const { name, email, password } = createUserBodySchema.parse(req.body);
-
-    const userExists = await prisma.user.findUnique({
-      where: { email },
-    });
-
-    if (userExists) {
-      throw new AppError("User already exists", 409);
-    }
-
-    const passwordHash = await hash(password, 10); // hashe da senha
-
-    const user = await prisma.user.create({
-      data: {
-        name,
-        email,
-        password: passwordHash,
-      },
-    });
-
-    const response = {
-      success: true,
-      message: "User created successfully",
-      data: {
-        ...user,
-        password: undefined, // nao retornar a senha no response
-        updatedAt: undefined, // nao retornar campo updatedat
-      },
-    };
-
-    res.status(201).json(response);
-  } catch (error) {
-    // Tratamento específico para erros de validação do Zod
     if (error instanceof z.ZodError) {
       const validationErrors = error.errors.map((err) => ({
         field: err.path.join("."),
